Experiment 1 from Mason et al. (2024)

This dataset is from Experiment 1 of Mason et al. (2024). (Rare and extreme outcomes in risky choice). Data is publicly available on OSF: https://osf.io/hy3q4/. We performed basic cleaning to meet our package needs.

Format

A data frame with 45000 rows and 11 columns:

Subject

Subject ID, an integer (16 to 144)

Block

Block number, an integer (1 to 6)

Trial

Trial number, an integer (1 to 60)

L_choice

Left choice, A = 100% gain 4, B = 90% gain 0 and 10% gain 40, C = 100% lose 4, D = 90% lose 0 and 10% lose 40.

R_choice

Right choice, A = 100% gain 4, B = 90% gain 0 and 10% gain 40, C = 100% lose 4, D = 90% lose 0 and 10% lose 40.

L_reward

Reward associated with the left choice.

R_reward

Reward associated with the right choice.

Sub_Choose

The chosen option, either L_choice or R_choice.

Frame

Type of frame, "Gain", "Loss", "Catch".

NetWorth

The participant's net worth at the end of each trial.

RT

The participant's reaction time (in milliseconds) for each trial.

Examples

  # Load the Mason_2024_Exp1 dataset
  data(Mason_2024_Exp1)
  head(Mason_2024_Exp1)
#>   Subject Block Trial Frame L_choice R_choice L_reward R_reward Sub_Choose
#> 1      16     1     1  Loss        C        D       -4        0          C
#> 2      16     1     2  Loss        D        C        0       -4          D
#> 3      16     1     3  Loss        C        D       -4      -40          C
#> 4      16     1     4  Loss        C        D       -4        0          D
#> 5      16     1     5  Gain        B        A        0        4          B
#> 6      16     1     6 Catch        C        B       -4        0          C
#>   NetWorth   RT
#> 1       -4 2318
#> 2       -4 3029
#> 3       -8 1897
#> 4       -8 1067
#> 5       -8 1814
#> 6      -12 1362